requirements.txt
Django==6.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=10.4.0
qrcode==7.4.2
twilio==8.10.0
psycopg2-binary>=2.9.11
python-decouple==3.8
firebase-admin>=6.0.0
